AG-13343 Add RovingTabContainerWidget to re-use keynav logic
